Move buffer types into private interfaces
